+++ LethalAPI.Monsters/NuGet.config
@@ -1,0 +1,6 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<configuration>
+	<packageSources>
+		<add key="BepInEx" value="https://nuget.bepinex.dev/v3/index.json" />
+	</packageSources>
+</configuration>
